Remove k6ThrowOnce and friends
When you run tests only once, there is no problem, however if you run
them multiple times using the count flag, k6ThrowOnce kicks in and
cause k6ThrowUnsafe not to run. This happens because the Go test runner
compiles the test binary only once and uses it to run the test in the
same process.

I don't know why but somehow we or Goja have fixed the issue in k6Throw
and now this commit entirely removes the k6Throw synchronization.

I tested it several times and I didn't observe any issues.

Fixes: #218