Move the allocator pkg into chromium pkg
The allocator pkg was there to kill the child process when its parent
also dies. However, it had a single function with a couple of lines of
code just to that. A little copying is better than a little dependency.
So I think we can use the same code for other browsers when it becomes
a reality (far away for now).

This commit removes the Allocator interface. A preemptive interface is
unnecessary and a premature optimization at this point and we don't
need an interface for doing this for other browsers as well. We can get
benefit from implicit interfaces which allows to extend the design much
more easily. There is no need to follow Java practices.