Fix `WaitForNavigation` within the same document (#247)
* Fix WaitForNavigation within the same document

Previously WaitForNavigation assumed that a LifecycleEvent would always
be fired, but this is not the case for navigation within the same
document (e.g. via anchor links or the History API), so in those cases
the call would timeout after 30s.

The fix simply checks if we received a new document, otherwise it skips
waiting for the LifecycleEvent. Even if that wait didn't time out, it
would've failed with a nil pointer panic accessing event.newDocument.request.

Closes #226

* Move static HTML to standalone file

Resolves
- https://github.com/grafana/xk6-browser/pull/247#discussion_r810969489
- https://github.com/grafana/xk6-browser/pull/247#discussion_r810974588

* Send empty struct instead of closing done channel

Resolves https://github.com/grafana/xk6-browser/pull/247#discussion_r810990288

* Use Page.click() instead of ElementHandle.Click()

Resolves https://github.com/grafana/xk6-browser/pull/247#discussion_r811003993