Fix curl command while installing wix tools in CI
-O uses the remote name, so the 'wix311-binaries.zip' argument wasn't
being used. We could keep -O but it's more reliable to specify our own
filename.

But this is a fix mainly to follow redirects with -L, since the failure
we were seeing is:

curl: Remote file name has no length!
curl: (23) Failed writing received data to disk/application

Which is because the first response is a redirect.

Also adds -s to hide the progress bar.