Fix checkHitTarget bug (#171)
When we run the script in the issue #171, it results in "waiting did
not result in any nodes error". It's because we think the element we're
looking for is clickable when it's not (another element gets above the
target element and makes it not clickable in practice).

checkHitTargetAt now returns with an error if the target element is
obsucured by another element; instead of waiting for the target element
to become clickable.

This fix is needed before improving the scrolling. Otherwise, we would
think that we properly scroll to the target element but cannot click on
it somehow!

Other notes:
-----------
* cdppage.GetLayoutMetrics now uses CSS layout metric when calculating
  the client inner width and height (viewport) instead of depending on
  the first layout parameter (deprecated and was returning incorrect
  results).

* We now count the target element clickable if its surface area is
  visible >0.99px. It was 1.01px before.