Add TestElementHandleClickConcealedLink
The test makes a long list of items, some are hidden, some are visible.
But the target element that we want to click is hidden. So we need to
scroll into it.

However, the default scrolling mechanism (that has no scroll options)
cannot do that.

This test verifies that we can scroll into the concealed and clickable
elements.