Add scroll retry to pointer action
getElementHandlePointerActionFn:
If scrollRectIntoViewIfNeeded is enough, just uses it if no other
scrolling options are given. This is the default behavior.

If not, then it scrolls into the element using the JS-native
scrollIntoView function. This one does not exist on CDP, so we need to
use JS evaluation here.

retry:
First tries to scroll into the clickable element with the default method
(no scrolling options). If it fails, then starts retrying the pointer
action with other scrolling options until it can find a clickable
element.