Upgrade to new JS extension API
This drops all usage of context-based getter functions (removed from k6 in [1])
in favor of methods attached to the VU instance passed to the extension.
We attached this VU instance to the internal context in #268 to keep a
similar API, but in the future this should be refactored to avoid the
use of context.

[1]: https://github.com/grafana/k6/commit/64f17c7e8dbdf283ed4729bf0c2967eb278071ad