Refactor test mock VU initialization
This centralizes where the goja.Runtime, k6lib.State, etc. are created by
adding a newMockVU() helper function to DRY things a bit. Unfortunately the
function is duplicated in both common and tests packages since it wasn't
possible to have it in a single place because of import cycles (unless we want
to make it part of non-test common code, which we preferably shouldn't do).