Fix negative metric emission
This issue happened when receiving `Page.lifecycleEvent`s for different
frames which overwrote the global `FrameSession.initTime` value on which
the metric value calculation was being done.

The fix is simple: store the `init` time on `Frame` instead and
calculate the elapsed time per frame, since the metric emission is done
per frame anyway.

This also does some refactoring to DRY things.

Closes #191