Remove goja.Runtime usage from ExecutionContext.eval
This was causing data races if the runtime was also being used by the
event loop goroutine, since goja is not thread-safe. It also moves
towards removing the dependency on goja from internal methods, which is
what we're aiming for. A good rule of thumb to follow from now on is to
never use goja on unexported methods, and only on those that are
directly exposed to JS.

Part of #271