Fix zombie browser processes
I noticed that we often don't GC browser processes. Sometimes, this
may affects test runs by reducing the available resources on a machine
(especially GH CI machines)—hence even may cause false test failures.

The problem happens because we only listen for EventConnectionClose in
Browser. If we ever receive this event, we close the lostConnection
channel via browserProc.didLoseConnection(). However, we might never
receive this event because of context cancelation (and it happens pretty
often AFAICS).

This PR uses a better approach would be to close the browser process
whenever the context is canceled.

Fixes: #305