Add locator.click
Clicks on an element with strict mode on. In the test, that is why it
expects a panic when the test tries to select multiple link elements.
The test also verifies whether the locator can click on an element using
the saved selector in the locator.

Pitfalls: Although locator.Click forwards the click call to frame.Click,
I wished there was a way to set the incoming goja value's strict
property to true, instead of parsing the options in locator.Click. Do
let me know if you know a way to do so.