Add the ability to combine keyboard presses.
This change enables the `Shift` to alter the case of letters if the
special `Key*` is used (where * is A-z). The shift key is only used
when working with certain keyboard methods, and `Type` is not one of
them. The change follows the same behaviour as was found in PlayWright.
The separator is `+`, so "Shift+KeyA" would be split into "Shift" and
"KeyA".

Closes: https://github.com/grafana/xk6-browser/issues/285