Not sink the global metric
The global metric is controlled by the k6's ingester output,
so the sink operation on the global metric in the sample is already executed.

Use a local metric in the metricsStorage for executing a local and dedicated sink
in the extension.

It fixes the data race where both the extension and the k6 ingester
calls Add and sink for the same metrics.