Refactor logger into its own package
Logger is a common process but it should live in its own package, that
is dedicated to all things logging. This will also allow us to easily
import the logger anywhere, helps avoid cyclic imports, it gives us a
cleaner api, and could allow us to migrate to other logging modules if
we needed.