Fix cannot fill an input text box
The Fill method cannot set the value property of an input text box
because it cannot detect that it's an input text box, and returns
prematurely (early).

This commit fixes the issue in the injected script by removing the
kDateTypes check, and makes the if clause under it applicable both for
kDateTypes and kTextInputTypes. We check for kDateTypes and
kTextInputTypes already in the fill method; so there is no need to check
it again.

It also adds a test for reproducing the issue.

Fixes #382