Use require a lot more than assert in js tests (#2564)
The js tests had a bunch of places where they were using `assert` instead of `require`.

Mostly cases where the test should've not continued such as expecting error or no error. In some places `if assert....` was used which also is completely not needed.

I have kept most usage of `assert` where they will report multiple (and even singular failures sometimes), but I would argue in most cases just using `require` for everything is likely enough. 

There are also a few places where I've added `requires` that were previously missing.

Finally, there are few places that could've been left as `assert` but I decided to go for `require` in order for there to be no `assert.Error` for example. As I would argue this should always be with require.

Some of those lead to *fewer* reported failures as previously an `if assert ...` was followed by more code that could run even if that assert is a failure. But I would argue in all of those cases that is unlikely, and even if that happens this particular failures will likely not help debug the original problem.