Add to showcase how currently options is accessible everywhere
This is also true for anything defined in the main module exported or
not. It is just much easier to showcase with options particularly.

This IMO should never have been the case and should be fixed.

Additionally ESM in practice will "fix" this as modules all have their
own scopes. So it makes sense to maybe break this before we add ESM
support