Fix max duration calculations with multiple 0 vus stages at end
In b725f03b we fixed the calculation of the stages, unfortunately we
didn't notice that the max duration for the test was wrongly calculated.

This also tries to fix cases where graceful* will extend the test more
then needed. Previously the code did not take into account that multiple
0 VUs stages can mean that the VUs are long ago stopped before
gracefulStop (for example) needs to be taken into account.