Refactor locator test to check for goja bool
Adds two helper methods to the test browser to reduce mistakes and turn
around time in code reviews. So that we will never forget to check the
validity of converted goja values. Other helper methods can be added in
the future if we need them. Another side benefit is that it reduces the
noise in the tests and let us see the main concerns of a test.