Improve context errors on panic
Instead of saying "context deadline exceeded" we slightly improve the
error and say "timed out". There are ways to include "timed out after
1s", however, that would complicate code. Let me know if you want me to
figure it out.

Another pitfall is that we look for the last argument of the Panic
function to get the error. Since we commonly add the error as the last
argument, this should make sense. I searched for the entire codebase and
couldn't find that does otherwise.

ERRO[0005] launching browser: getting DevTools URL: timed out
        at reflect.methodValueCall (native)
        at file:///Users/inanc/grafana/xk6-browser/examples/browser_on.js:6:14(13)
        at native  executor=per-vu-iterations scenario=default source=stacktrace