output/json: Fix json data race
The previous implementation makes a copy of the Metric struct
where it generates a data race on the Observed field that is written
from the metrics.Engine.

The current implementation reduces the fields copied resolving
the race on Observed. Observed field is not useful for the JSON output.

Thresholds shouldn't be racy because the output mostly read
only the "immutable" values.

The Tainted shouldn't be racy because the metrics are flushed only at
the begin. However there isn't a guarantee for it, so based on the fact,
it isn't useful because it will never be different from null, we will
drop it.

The performance seems not impacted, it is aligned with the
result from a5146cf7099cfc5b4ec8938c667c507ce373aff9.

BenchmarkFlushMetrics-8   	      61	  16767247 ns/op	 2526610 B/op	   20067 allocs/op