Refactor test so that we're checking the promise
We just want to make sure that the promises actually complete and are
fulfilled before we end the test, so we're checking the state of the
promise in `tb.await` as well as after `tb.await`. There's a chance
that the action has already completed before the first state check, but
the resolve/rejection only occurs `tb.await` returns.

Resolves: https://github.com/grafana/xk6-browser/pull/467#discussion_r932044728