Fix the data race with injectedScript
A data race can occur when two tests perform an action on the same page
but under two different threads. It occurs in ExecutionContext with its
injectedScript parameter. Since we call
ExecutionContext.getInjectedScript() in ElementHandle.evalWithScript(),
it can be called in parallel from different goroutines, so a mutex is
a good fix for this.

Closes: https://github.com/grafana/xk6-browser/issues/426