Setup waitForNavigation waiters outside of the promise
This (mostly) ensures that the waiters are started before
`waitForNavigation()` returns, and that the promise handler simply waits
for them. It reduces the chances of a race condition where e.g. a
navigation event produced by a `click()` is received before the waiters
are setup, leading to 30s timeouts. This used to happen even when
Promise.all() was used.