Fix press in classes to work with combo keys (#522)
* Fix press in classes to work with combo keys

ElementHandle, Page, Locator and Frame all have the Press method. In a
recent fix (#326) combo keys with the Shift modifier key was fixed for
the Keyboard Press method. The Press methods for the other classes
should have worked with combo ley presses, but didn't. This fix will
ensure that combo key presses work for all of them.

Closes: https://github.com/grafana/xk6-browser/issues/521

* Apply suggestions from code review

Co-authored-by: Ivan Mirić <ivan.miric@grafana.com>

Co-authored-by: Ivan Mirić <ivan.miric@grafana.com>