Close all sessions instead of the connection after sending Browser.close
Browser.close will trigger several things from the browser:
- Sending of Target.detachedFromTarget events for each page.
- An exit of the process without cleanly closing the WS connection. I.e.
  we will receive an `unexpected EOF` error from websocket.ReadMessage().

We need to wait for all detachedFromTarget events to be processed, which
is achieved by waiting for <-b.browserProc.processDone. But we won't
bother closing the WS connection from our side, since a) it could have
already been closed by the time we try, and b) more importantly, it
might cause us to prematurely close it and miss any detachedFromTarget
events. So, we just don't bother with closing it and expect the browser
exit to happen, and we'll catch and ignore the `unexpected EOF` error if
we sent out Browser.close.