Dont lock the event loop once a promise rejection is left unhandled
Previously when a promise is rejected but *not* handled, the event loop
would end the current iteration and then let a new one start.

Unfortuantely it also didn't clear the failure so the next time it will
again *fail* the iteration as it still thinks a promise has been
rejected without a handler.

Heavily related to https://github.com/grafana/xk6-browser/issues/510