Add a queue per handler to keep events in order
Events are sometimes out of order due to use using a goroutine to
publish events to the handler concurrently, so there's no guarantee
that the goroutine will run in the order they're created. To overcome
this, a queue has been added per handler to synchronise all publishes
of events to that handler. This should ensure that the ordering of
events consumed by the handler matches what is emitted. We can't remove
the goroutine in emitTo as this deadlocks due to handlers consuming and
emitting events.

Closes: https://github.com/grafana/xk6-browser/issues/553