lib,js: Set always the Tag URL equals to tag Name
After the introduction of Atlas, we decided to keep `url` as a normal tag
but it always set to have the same value of the `name` tag.
In this way, if the `url` contains high-cardinality values it wouldn't
affect the system.
This is the same thing that currently the cloud output does.