Fix networkIdle waitUntil option for load
When a site that is navigated to emits a DOMContentLoaded well before
load, this is a signal that we're waiting on assets to load over the
network. In such a case (e.g. navigating to google.com) when we goto
with waitUntil = networkIdle, due to how we handle such lifecycle
events, we end up in a situation where the networkIdle event is
ignored. This change ensure that we action on these events too.

Partly fixes: https://github.com/orgs/grafana/projects/80/views/14