Fix clearLifecycle for goto
When we navigate to a page, lifecycle events were being stored. If a
new navigation occurred, the previous events would be used and fired
off to the internal handles. This is incorrect as it means that
WaitUntil will fail to wait for the correct event after the initial
navigation to a page.

Partly fixes: https://github.com/orgs/grafana/projects/80/views/14