Fix WaitUntil for long running network requests
Some websites may have many network requests which are performed after
the page has fully loaded and we receive a FrameStoppedLoading. For
now we want to wait for the networkIdle event, and in the future if we
need, we should add a timer for very chatty websites where we may
never see a networkIdle event.

Partly fixes: https://github.com/orgs/grafana/projects/80/views/14