Fix race condition with frame.inflightRequests
When logging the frame's inflightRequests, there's a chance that the
inflightRequests are being accessed by another goroutine.
inflightRequests is already protected with a mutex within Frame, so
we should use the same protection when logging the requests. In this
change we add a getInflightRequest which makes a clone of the original
(not a deep clone though).
There's a chance that was it being logged is already old if the
inflightRequests is also being written to at the same time.