Remove neworkidle timer as its not needed
This timer is no longer needed since the fixes in PR #555 and #597.
These PRs helped to ensure that we receive the CDP events in order
and that we correctly handle them when we do receive them. The
networkidle timer was in place for a situation where we might not
receive the networkidle lifecycle event from the browser via CDP.
This could still occur, which would mean that there's a bug outside
of our direct control. When this happens we should find a better
solution.