Remove the need to recalcLifecycle for sub frame
Each frame (main and the subs) receive their own lifecycle events in
the correct order (so the sub frames first, and then the root/main
frame). The reason for having this code in the first place I believe
is due to out of order CDP event, which has now been fixed in PR #555.

After extensive stress testing I can confirm that we no longer need to
perform the recalculateLifecycle in a recursive way. So each frame
should be in charge of recalculating its own lifecycle.

There maybe a need to do this in the future if the browser changes its
behaviour on how it deals with lifecycle events, but at this point we
should not consider adding such a recursive call.