Refactor the emitting of the lifecycle event
We're only ever going to fire the newly received CDP lifecycle event
so there's no need for the for loop to go over all the events that
have already fired and ignore the ones that have already been emitted
to internal handlers.

This should lead us to refactoring out the subtreeLifecycleEvents
as there's no need for it now, it's just a duplicate of
lifecycleEvents.