js: make certain only one copy of the exports is used per each module
This is inline with ESM and works for source code modules, but not for
modules written in go as teh code was never made to work with it.

This also lets (as the test suggests) change what `k6/http`.get is
across the VU just as it should be possible.