Add custom errors
The Web Crypto API specification has a pretty loose set of expectations
towards errors. However, making our implementation pass the Web
Platform Tests outlined that they're expected to be in a quite specific
format.

In order to provide a good support for errors in the Web Crypto API,
we define its expected errors as custom errors.