js: refactor how modules are loaded
This refactor tries to simplify the implementation of `require` and
connected code by splitting it heavily into different parts.

These changes are similar to what will be needed for native ESM support
as shown in https://github.com/grafana/k6/pull/2563 , but without any of
the native parts and without doing anything that isn't currently needed.
This will hopefully make ESM PR much smaller and less intrusive.

This includes still keeping the wrong relativity of `require` as
explained in https://github.com/grafana/k6/issues/2674.

It also tries to simplify connected code, but due to this being very
sensitive code and the changes already being quite big, this is done
only to an extent.

The lack of new tests is mostly due to there not being really any new
code and the tests that were created along these changes already being
merged months ago with https://github.com/grafana/k6/pull/2782.

Future changes will try to address the above as well as potentially
moving the whole module types and logic in separate package to be reused
in tests.