Wait for metrics handling to finish even when there was an error
Previous to this, setup() and teardown() exceptions or test.abort() calls would have made k6 exit prematurely, without waiting for the metrics processing to finish. This also caused a data race in integration tests and undefined behavior in general. The handleSummary() function was also called only when test.abort() was used, but not when there was an error in `setup()` and `teardown()`.

The commit also adds and improves a bunch of integration tests, particularly when it comes to validating aborted tests and --linger.