Fix k6 Hosts option incompatibility
We first update the k6 dependency (see go.mod and go.sum). Then we fix
the breaking code.

All tests pass. But, this solution is not an ideal one. As we are
marshaling and unmarshaling to grab the hosts (see
chromium/browser_type.go).

Since the whole API has changed, I needed to fix tests with the new API
as well.

A proper solution would be to use the new API in our network code,
instead of passing hosts to block to Chromium.

PS: Sorry, I couldn't keep this PR shorter. The problem was CI would
break if I sent each change as separate commits.