Fix remote frame detach logic
When a website contains an iframe where its src is not from the same
origin of the current frame (e.g. embedded youtube video), we get
two events from chrome. Firstly we get a detach, and then an attach.
The detach we need to handle differently and leave the attach as is.

In the detach, instead of removing the frame that the detach event
points to, we need to keep it but clean up any child frames it may
reference. The detach (with swap reason) signals that a remote frame
(that will be rendered by a new process) will be attached to the
referenced frame. Later we will receive an attach frame event that

Later we will receive an attach frame event that references the same
frame as the detach as its parent, and this will be the new remote
child frame.

Fixes: #669