Add TraceID construct to generate trace ids
This commit adapts what was already developed in the
xk6-distributed-tracing extension. It introduces the TraceID struct,
constructor, and encoding method.

Note that in this commit we kept the logic relative to prefixing the
trace ids with k6 specific information. That is for no other good
reason than the apparent absence of bad reasons to keep it. The ability
to find out if a trace_id was generated by k6 sounded potentially useful
to justify keeping it.