Implement and expose a tracing.Client
This commit implements the tracing Client object. The object exposes a
similar API as the http module: request, get, head, etc. It behaves
in a similar fashion, while adding tracing context information to the
requests it performs, and ensuring the trace id is emitted as metadata
on the output's data points.

It is a convenient way of explicitly perform instrumented HTTP calls,
while keeping the standard k6 http module untouched, and allows to have
instrumented and non-instrumented calls side by side in a single script.

The client wraps the http module logic in a way that allows it
to fill each request's params argument on the fly with the tracing
headers generated by the selected propagator. It also ensures that
the related output metadata are set before and after the http request.

Note that this commit does not implement support for the batch and
request methods for technical reasons justifying more research and
an implementation in a separate context/commit.