Map [Page|Frame].Click and fix tests
This moves the promise handling to the mapping layer.

Some flakiness from tests are expected at this stage. To reduce
flakiness, this PR introduced intermingling of non-async and async code
in tests. For example: TestWaitForFrameNavigationWithinDocument. It
wraps the click in a promise to run WaitForNavigation together with the
click. We will do this using goroutines when we remove the promisified
WaitForNavigatin soon (probably, through a helper).