Map [Page|Frame].WaitForFunction promises and fix tests (#720)
* Move waitForFunction close to where it's used

Go code should be written top to bottom readable.

* Map [Page|Frame].WaitForFunction and fix tests

This moves the promise handling to the mapping layer. Ideally, this
tests should be written by the JavaScript side. See issue #675 for
details. The reason is, they are completely concerned about the JS-side
of things and the Go tests become futile, complicated and hard to
understand.