Implement and experimental tracing module
* Implement trace-id generation logic

This commit implements trace-id generation based on the recommendations
of the W3C specification.

A lot of the code and logic present in this
commit are inherited and refactored from the xk6-distributed-tracing
project.

* Implement trace context propagators

This commit implements the trace-context propagation logic, and allows
for the creation of trace context headers in the W3C, B3 (openzipkin),
and Jaeger formats.

Note that a lot of the code and logic present in this commit are
inherited and refactored from the xk6-distributed-tracing project.

* Register a k6/experimental/tracing

Although we do not expose any of user-facing logic at this point, we
register the "k6/experimental/tracing" module in k6 to prepare for
further implementation of public APIs

* Define a tracing Client

This commit defines a Client type which exposes an API similar to the
HTTP module, and allows users to perform HTTP requests embedding a
trace-context, and attaching their trace_id as an output metadata to
their HTTP-related samples.

It uses the HTTP module's request function under the hood, and wraps it
with some "hook" logic to ensure the expected headers are present, and
that the expected output metadata are emitted.

The tracing client exposes an option-set allowing to use the three
available propagation format. Other options such as sampling, and
baggage are present for forward-compatibility, but will be implemented
at a later point in time.

* Expose the tracing client publicly as k6/experimental/tracing.Client

This commit exposes the Client constructor publicly as part of the
k6/experimental/tracing module. From this point forward users will be
able to instantiate the Client, and perform instrumented HTTP requests
using it.

This commit also adds a bunch of integration tests covering the expected
behavior of the module's API.

* fixup! return error instead of common.Throw in Client.Request

* fixup! make counters in tests atomic

* Expose the instrumentHTTP function as part of the module's public API

* Add integration tests covering the instrumentHTTP function

* Update cmd/tests/tracing_module_test.go

Co-authored-by: Ivan <2103732+codebien@users.noreply.github.com>

* Update cmd/tests/tracing_module_test.go

Co-authored-by: Ivan <2103732+codebien@users.noreply.github.com>

* Add unit tests for the instrumentHTTP function

* Remove unnecessary integration tests

* Add integration test illustrating instrumentHTTP in multi-file setup

* Move unit test setup in dedicated struct

* Adjust test ensuring instrumenting function from other module works

* Move tracing-client example in the correct samples folder

Co-authored-by: Ivan <2103732+codebien@users.noreply.github.com>