Make Propagators support sampling by implementing the Sampler interface
This commit defines a SamplerPropagator interface, which established
the contract for a propagator which bases its sampling flag on a
Sampler implementation.

It also makes the concrete implementations of Propagators implement
the Sampler interface, and ensure the flags field of the generated
trace context headers is set based on the Sampler's result.