Provide default logger when given is nil
If the given inner logger for our log implementation is nil,
provide a new logrus logger by default. This avoids a possible panic
when checking for the log level in the Logf method as well as allows us
to remove the particular logging action using Printf when wrapped logger
was nil.

Resolves: https://github.com/grafana/xk6-browser/pull/734#discussion_r1089171642