Fix panic with setInterval in k6/ws
The panic happens when interval between 0 and 1 is given.

The code that tested whether the interval was > 0 correctly converted to
floats first and then back to time. And then checked it in order to
prevent this very panic.

Unfortunately it then did the same calculation again but this time by
converting the interval directly to int before multiplying by
milliseconds. Which ends up rounding it to 0. Making the whole result
zero and leading to time.NewTicker to panic.

This was already done correctly for setTimeout when the same panic was
found there and the wrong conversation was added for setInterval.

Previous commit df3ad1bd