js: Cache scripts by absolute not relative paths
This fixes a regression introduced in 2753196ab where in all imports are
cached on the specifier that is provided for them. This is fine for k6
modules which happen to always have the same absolute names.

But source modules can have relative paths that resolve to different
absolute ones depending on where they are imported from.

Fixes #2902