k6.check: Fix returned values when iteration is ending
Before this check was only returning `false` if it also could emit it,
and true otherwise.

This should rarely be a problem as the same context that is being
checked here is the one that interrupts the VM - so no code will be able
to run after this.

Unfortunately the code checking if it should be emitted is racing with
the one that interrupts the VM. So it is possible for the VM to still
not be interrupted when a `check` returns a wrong value. It is even
possible for more code to be run before the interrupt is actually
called.

The code still checks the context as this also updates the internal
check structure and we don't want to that if the context is done.
The above should be changed with #2869

Fixes #2912