Fix multiple browser closing on panic
When a user script launches two browsers in the same tests and one of
them panics, we can now kill both of them.

Earlier, we used to store the browser PIDs in the context and that was
problematic because we couldn't kill all the browser processes. We can
now remove doing that.

Since RootModule stores all the pids, we can now get and kill them all
when k6ext.Panic is called.

In tests, there is no need to kill the processes as the testBrowser sets
them to be killed if they closed on panic.