Add BrowserProcessMeta abstraction
Adds a new component BrowserProcessMeta in order to abstract the
handling of the browser process handle and the browser data directory.
This is defined as an interface that will support implementations for a
local browser process and for a remote browser process and therefore
minimizing changes in the rest of the implementation, as well as
protecting the process and data dir properties from possible NPE errors.