Remove BrowserType logger
Removes the logger instance held by BrowserType as it's unnecessary.
Instead a local logger is created and passed along during the launch
process. Also removes the unnecessary BrowserProcess.AttachLogger method
as the logger was already passed to the constructor and could be set
from there.
This will allow to not tight the loggers used for BrowserType.Launch and
BrowserType.Connect methods.