Drop a sample that uses a removed REST API
This sample is 6 years old and uses a version of k6 cloud (at the time
Load Impact) that predates k6 itself and has been removed for several
years now.

The sample while more on the complex side is not very useful as it can't
be run.

There are also now a lot more example in the wild of complex scripts
then 6 years ago.