Fix active VU reporting by arrival-rate executors
This is a somewhat quick and dirty fix of #1977, where instead of
changing the global active VUs when the VU is activated, we change it
when the VU is actually running. This synchronizes it with the running
count of the internal activeVUPool, and reports the correct value
everywhere the ExecutionState is used (the real-time total active VU
count while the test is running, in all outputs, etc.).