Add support for reporting DNS resolution times
The time it takes to resolve hostnames is missing from the reported
metrcs. Add support for recording that information thru the httptrace
DNSStart and DNSDone hooks. This requires changing the signature of the
netext.MultiResolver type because nettrace depends on having the hooks
passed in a structure in the request context (they are wired up by
httptrace).

Signed-off-by: Marcelo E. Magallon <marcelo.magallon@grafana.com>