Make TraceID randomness source selectable
In order to be able to test how the client updates VUs metadata with
a trace_id, we need to be able to predictably control the generated
trace ids.

Because the traceID encodes itself, we set a private an `io.Reader`
randomness source attribute on it that will be used during encoding.
That way we can for instance pass a reader that always produces the
same value in the context of tests.