Fix page.close so that it closes the current page
This will close the current page instead of closing the browserContext
that the page is associated to when page.close is called.

page.close uses target.CloseTarget. This CDP command initiates a
cleanup process in the browser, and a side effect of this is that
the browser will run target.EventDetachedFromTarget. When k6-browser
sees this event, it immediately closes the session connection which
the close command was performed on, and so this results in a context
closed error. This is racey though and other times the target.Close-
Target command will return before we see the target.EventDetachedFrom-
Target event.

Closes: https://github.com/grafana/xk6-browser/issues/833