Add the downloaded web vital js library
We're going to bundle the web vital js library from the internet
instead of downloading them on each test run. This will ensure that
users who run tests in a locked down environment can still work with
web vitals, and it will avoid any CORS related issues. The downside
is that we will need to manually update this library, and users who
do not update to latest k6 binary will be working with an outdated
version of the library.