Add implementation to page.evaluateOnNewDocument
This will evaluate the script on the current page. BrowserContext
relies on page.evaluateOnNewDocument when a new init script is added
so that they are applied to all existing pages.

For now, the preference is not to work with the worldName param in the
AddScriptToEvaluateOnNewDocument CDP command. For now we don't want
the script to be isolated to a particular named world.

The script identifier that is returned from a successful CDP command
to add the init script, is ignored. When we need it (e.g. for removing
init scripts) we should use it then.

Co-authored-by: Daniel Jiménez <daniel.jimenez@grafana.com>