Refactor NewBrowserContext to return error
By refactoring this one method, we need to follow the chain up so that
other methods that work with NewBrowserContext also return an error.
We need NewBrowserContext to return an error so that we can validate
the web vital init script ,which is currently being added to
evaluateOnNewDocumentSources without any validation (change to come in
the next commit).