Simplify the CryptoKey type
The pre-existing CryptoKey type was intended as a contract-type
declaration of type matching the specification's. It was not bound to
any specific implementation, and as a result, not battle tested.

This commit realigns its definition with the real-world use-case
brought by the `subtle.crypto.generateKey` operation's implementation.

The two main changes are the switch from a generic type to a `any` type
for the `CryptoKey.handle` and `Cryptoattribute. This attribute is not exposed to
the user, and is heavily contextual. We will in practice know from the
context what type it is supposed to be castable to. The second change
is to define a dedicated generic type for the algorithm field. There
are a variety of Algorithm object types defined by the specs, and we
need to be able to distinguish and export them to dedicated objects
easily for them to be convinient to use.