Add the web vital metrics to the metrics registry
This adds the six web vitals, as well as the three ratings for each.
When a web vital metric is send from the JS code to k6-browser via
the binding, it will need to extract the metric name and the rating.

The rating can be good, needs-improvement or poor.